test(cli): split bundle tests
