+++ packages/core/test/cli/bundle/errors.ts
@@ -1,0 +1,8 @@
+import { describe, test, expect, runCli } from '../helpers';
+
+describe('CLI: bundle errors', () => {
+  test('존재하지 않는 entry → 에러', () => {
+    const { exitCode } = runCli(['--bundle', '/nonexistent/entry.ts']);
+    expect(exitCode).toBe(1);
+  });
+});
